# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "BISQ Support",
        "version": "0.1.0",
        "description": "BISQ Support API - A RAG-based support assistant with admin authentication.\n\n## Security\n\n### Transport Security (CKV_OPENAPI_20)\nProduction deployments use HTTPS exclusively via nginx reverse proxy configuration.\nAPI keys are only transmitted over secure channels in production.\nLocal development uses HTTP for debugging convenience.\n\n### Authentication Design (CKV_OPENAPI_4, CKV_OPENAPI_5)\nThis API uses **per-endpoint security** by design:\n- Public endpoints: /health, /chat, /feedback (no authentication)\n- Admin endpoints: /admin/* (require AdminApiKeyAuth header)\n- Exception: /admin/auth/logout is public (no authentication required)\n\nNo global security is defined because public and authenticated endpoints coexist.\n\n### Array Limits (CKV_OPENAPI_21)\nArray properties intentionally omit maxItems to allow flexible data volumes.\nActual limits are enforced by:\n- Application logic (e.g., FeedbackRequest.conversation_history maxItems=50)\n- Backend validation in FastAPI route handlers\n- Database constraints and query limits"
    },
    "paths": {
...
        "/admin/faqs": {
            "get": {
                "tags": [
                    "Admin FAQs"
                ],
                "summary": "Get All Faqs For Admin Route",
                "description": "Get FAQs for the admin interface with pagination and filtering support.",
                "operationId": "get_all_faqs_for_admin_route_admin_faqs_get",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Page Size"
                        }
                    },
                    {
                        "name": "search_text",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Search Text"
                        }
                    },
                    {
                        "name": "categories",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Categories"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Source"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FAQListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin FAQs"
                ],
                "summary": "Add New Faq Route",
                "description": "Add a new FAQ.",
                "operationId": "add_new_faq_route_admin_faqs_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FAQItem"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FAQIdentifiedItem"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            }
        },
        "/admin/faqs/{faq_id}": {
            "put": {
                "tags": [
                    "Admin FAQs"
                ],
                "summary": "Update Existing Faq Route",
                "description": "Update an existing FAQ by its ID.",
                "operationId": "update_existing_faq_route_admin_faqs__faq_id__put",
                "parameters": [
                    {
                        "name": "faq_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Faq Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FAQItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FAQIdentifiedItem"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin FAQs"
                ],
                "summary": "Delete Existing Faq Route",
                "description": "Delete an existing FAQ by its ID.",
                "operationId": "delete_existing_faq_route_admin_faqs__faq_id__delete",
                "parameters": [
                    {
                        "name": "faq_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Faq Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FAQIdentifiedItem": {
                "properties": {
                    "question": {
                        "type": "string",
                        "title": "Question"
                    },
                    "answer": {
                        "type": "string",
                        "title": "Answer"
                    },
                    "category": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Category",
                        "default": "General"
                    },
                    "source": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Source",
                        "default": "Manual"
                    },
                    "id": {
                        "type": "string",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "question",
                    "answer",
                    "id"
                ],
                "title": "FAQIdentifiedItem"
            },
            "FAQItem": {
                "properties": {
                    "question": {
                        "type": "string",
                        "title": "Question"
                    },
                    "answer": {
                        "type": "string",
                        "title": "Answer"
                    },
                    "category": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Category",
                        "default": "General"
                    },
                    "source": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Source",
                        "default": "Manual"
                    }
                },
                "type": "object",
                "required": [
                    "question",
                    "answer"
                ],
                "title": "FAQItem"
            },
            "FAQListResponse": {
                "properties": {
                    "faqs": {
                        "items": {
                            "$ref": "#/components/schemas/FAQIdentifiedItem"
                        },
                        "type": "array",
                        "title": "Faqs"
                    },
                    "total_count": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Total Count"
                    },
                    "page": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "faqs",
                    "total_count",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "FAQListResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "AdminApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer abcdef12345`"
            }
        }
    }
}